refactor #74 패키지 구조 변경 cf. #52
