+++ src/main/java/com/squadb/workassistantapi/reservation/domain/ReservationException.java
@@ -1,4 +1,4 @@
-package com.squadb.workassistantapi.domain.exceptions;
+package com.squadb.workassistantapi.reservation.domain;
 
 public class ReservationException extends RuntimeException {
