+++ src/main/java/com/squadb/workassistantapi/reservation/dto/CommonResponseDto.java
@@ -1,4 +1,4 @@
-package com.squadb.workassistantapi.web.controller.dto;
+package com.squadb.workassistantapi.reservation.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
