+++ src/main/java/com/squadb/workassistantapi/reservation/dto/PaginationResponseDto.java
@@ -1,4 +1,4 @@
-package com.squadb.workassistantapi.web.controller.dto;
+package com.squadb.workassistantapi.reservation.dto;
 
 import lombok.Builder;
