+++ src/main/java/com/squadb/workassistantapi/reservation/dto/ReservationResponseDto.java
@@ -1,8 +1,8 @@
-package com.squadb.workassistantapi.web.controller.dto;
+package com.squadb.workassistantapi.reservation.dto;
 
 import com.squadb.workassistantapi.domain.BookCategory;
-import com.squadb.workassistantapi.domain.Reservation;
-import com.squadb.workassistantapi.domain.ReservationStatus;
+import com.squadb.workassistantapi.reservation.domain.Reservation;
+import com.squadb.workassistantapi.reservation.domain.ReservationStatus;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
