+++ src/main/java/com/squadb/workassistantapi/reservation/dto/ReservationSearchAllDto.java
@@ -1,6 +1,6 @@
-package com.squadb.workassistantapi.web.controller.dto;
+package com.squadb.workassistantapi.reservation.dto;
 
-import com.squadb.workassistantapi.domain.ReservationStatus;
+import com.squadb.workassistantapi.reservation.domain.ReservationStatus;
 import lombok.Getter;
 
 @Getter
