+++ src/main/java/com/squadb/workassistantapi/reservation/dto/ReservationSearchType.java
@@ -1,4 +1,4 @@
-package com.squadb.workassistantapi.web.controller.dto;
+package com.squadb.workassistantapi.reservation.dto;
 
 public enum ReservationSearchType {
     ALL,
