+++ src/main/java/com/squadb/workassistantapi/reservation/infrastrusture/config/QuerydslConfig.java
@@ -1,4 +1,4 @@
-package com.squadb.workassistantapi.domain;
+package com.squadb.workassistantapi.reservation.infrastrusture.config;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.springframework.context.annotation.Bean;
